+++ src/Synapse/Synapse/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 
 ## Upcoming Release
-* Updated `Update-AzSynapseSparkPool` to support new parameter [-ForeApplySettings]
+* Updated `Update-AzSynapseSparkPool` to support new parameter [-ForceApplySetting]
 
 ## Version 1.0.0
 * General availability of Az.Synapse